Game: Remove network.c and network.h
