+++ game/src/game/Graphics/CameraSystem.cpp
@@ -20,7 +20,6 @@
 
 #include "egolib/_math.h"
 
-#include "game/network.h"
 #include "game/mesh.h"
 #include "game/graphic.h"
 #include "game/game.h"
