+++ game/src/game/input.c
@@ -23,7 +23,6 @@
 
 #include "game/input.h"
 
-#include "game/network.h"
 #include "game/graphic.h"
 #include "game/egoboo.h"
